+++ changes.md
@@ -15,6 +15,7 @@
    On the long run will completely replace jquery-ui widgets
 10. In all sources specify "use strict" derictive, fix several connected errors.   
 11. Implement PFC, PLC and PMC draw options for TMultiGraph
+12. Use color palette store in the TCanvas
 
 
 ## Changes in 5.2.0